requirements.txt
# Core dependencies
ccxt>=4.4.91
numpy>=2.3.1
python-dotenv>=1.1.1

# Performance and monitoring
psutil>=6.1.0

# HTTP and networking
aiohttp>=3.12.13
requests>=2.32.4

# Data processing
pandas>=2.0.0

# Logging and utilities
colorama>=0.4.6

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.5.0
types-requests>=2.31.0

# Code formatting and linting
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0 